Made ENV/ARCHIVE_BUCKET mandatory
also removed option to change tag used to determine which buckets to
archive as `to-archive` makes sense and I don't want to make things
more confusing/configurable than they need to be.

We can alway make it an option in the future if/when this will need to
be an option.